asdf
asfd

+++ app/controllers/ManagerController.php
@@ -9,21 +9,16 @@ class ManagerController extends \BaseController {
 	public function changeNodeManager() {
 		$userId = Input::get('userId');
 		$nodeId = Input::get('nodeId');
-
 		$user = User::find($userId);
-
 		if(!$userId){
 			$msg = "관리자로 지정할 사용자를 선택하세요.";
 			$code = 0; 
 			return array('msg'=>$msg, 'code'=>$code);
 		}
-
 		$capsaicinGroup = Group::where('key','=','equip.capsaicin')->first();
 		$psGroup = Group::where('key','=','equip.ps.admin')->first();
-
 		// 유저가 타 관서 장비관리자였을 경우 원래 관서의 관리자 없애기
 		$prevNode = EqSupplyManagerNode::where('manager_id','=',$userId[0])->first();
-
 		if ($prevNode) {
 			$prevNode->manager_id = null;
 			if (!$prevNode->save()) {
@@ -34,11 +29,11 @@ class ManagerController extends \BaseController {
 		$havingEqGroups = Group::whereHas('users', function($q) use($userId) {
 			$q->where('id','=',$userId[0]);
 		})->get();
-
 		foreach ($havingEqGroups as $g) {
+			if (explode('.', $g->key)[0] == 'equip') {
 				$g->users()->detach($userId[0]);
+			}
 		};
-
 		// 선택한 노드의 관리자로 임명함
 		$node = EqSupplyManagerNode::find($nodeId);
 		if ($node->manager_id !== '') {
@@ -46,13 +41,10 @@ class ManagerController extends \BaseController {
 		}
 		$node->manager_id = $userId[0];
 		$node->last_manager_changed_date = date('Y-m-d H:i:s');
-
 		DB::beginTransaction();
-
 		if (!$node->save()) {
 			return App::abort(500);
 		}
-
 		// 관리자로 지정되면 장비관리자 그룹에 넣기
 		// 지방청 관리자면 지방청관리자 그룹에 넣기
 		
@@ -62,7 +54,6 @@ class ManagerController extends \BaseController {
 		} else {
 			User::find($userId[0])->groups()->attach($psGroup->id);
 		}
-
 		// 해당 노드의 전임 관리자는 그룹에서 빼기
 		if ($predecessorId) {
 			if ($node->type_code == 'D002') {
@@ -72,18 +63,20 @@ class ManagerController extends \BaseController {
 				$psGroup->users()->detach($predecessorId);
 			}
 		}
-
 		$user = User::find($userId[0]);
-
-
+		
+		if (!$user->activated) {
+			$user->activated = true;
+			if (!$user->save()) {
+				return App::abort(500);
+			}
+		}
+		
 		$msg = "관리자가 변경되었습니다.";
 		$code = 1;
-
 		DB::commit();
-
 		return array('msg'=>$msg, 'code'=>$code);
 	}
-
 	public function displayNodesSelectModal() {
 		$nodeId = Input::get('node_id');
 		return View::make('widget.node-selector', get_defined_vars());
